pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "satvlm"
version = "0.1.0"
description = "위성영상 분석을 위한 Vision-Language Model REST API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "conaonda"},
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-multipart>=0.0.9",
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "transformers>=4.47.0",
    "accelerate>=0.33.0",
    "peft>=0.12.0",
    "Pillow>=10.3.0",
    "numpy>=1.26.0",
    "rasterio>=1.3.10",
    "huggingface_hub>=0.24.0",
    "qwen-vl-utils>=0.0.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
